Initial commit for helm service
Please note that this is a squashed commit for all changes prior to
helm service. Initially, the project was conceived as a helm fork, but
later was moved to be a module with helm as a dependency. Squashing
all commits prior to helm service commits preserves exising helm
service commits.